src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp           == 3.11.16
aiosignal          ==1.3.2
annotated-types    ==0.7.0
antiorm            ==1.2.1
attrs              ==25.3.0
base58             ==2.1.1
bash               ==0.6
beautifulsoup4     ==4.13.3
bitarray           ==3.1.1
blinker            ==1.9.0
boto               ==2.49.0
certifi            ==2025.1.31
charset-normalizer ==3.4.1
chord              ==6.0.1
ckzg               ==2.1.0
click              ==8.1.8
colorama           ==0.4.6
cytoolz            ==1.0.1
db                 ==0.1.1
db-sqlite3         ==0.0.1
distlib            ==0.3.9
dnspython          ==2.7.0
email_validator    ==2.2.0
eth_abi            ==5.2.0
eth-account        ==0.13.6
eth-hash           ==0.7.1
eth-keyfile        ==0.8.1
eth-keys           ==0.6.1
eth-rlp            ==2.2.0
eth-typing         ==5.2.0
eth-utils          ==5.2.0
Flask              ==2.3.3
flask-cors         ==5.0.1
Flask-Login        ==0.6.3
fonttools          ==4.56.0
frozenlist         ==1.5.0
gamepy             ==0.1.5
graphviz           ==0.20.3
h11                ==0.14.0
hexbytes           ==1.3.0
hyperlink          ==21.0.0
idna               ==3.10
ipfshttpclient     ==0.7.0
itemadapter        ==0.11.0
itsdangerous       ==2.2.0
jam.py             ==5.4.136
Jinja2             ==3.1.6
jiter              ==0.9.0
jmespath           ==1.0.1
joblib             ==1.4.2
kiwisolver         ==1.4.8
legacy-cgi         ==2.6.2
lxml               ==5.3.1
MarkupSafe         ==3.0.2
mdurl              ==0.1.2
MouseInfo          ==0.1.3
msgspec            ==0.19.0
multiaddr          ==0.0.9
multidict          ==6.1.0
namex              ==0.0.8
netaddr            ==1.3.0
nose               ==1.3.7
npm                ==0.1.1
numpy              ==2.2.3
optional-django    ==0.1.0
packaging          ==24.2
parsimonious       ==0.10.0
php                ==1.2.1
pillow             ==11.1.0
pip                ==25.1
platformdirs       ==4.3.6
propcache          ==0.3.0
Protego            ==0.4.0
pyasn1             ==0.6.1
PyAutoGUI          ==0.9.54
pycparser          ==2.22
pycryptodome       ==3.21.0
pydantic           ==2.11.2
pydantic_core      ==2.33.1
PyDispatcher       ==2.0.7
pygame             ==2.6.1
PyGetWindow        ==0.0.9
Pygments           ==2.19.1
pymongo            ==4.11.2
PyMsgBox           ==1.0.9
pyotp              ==2.9.0
pyparsing          ==3.2.1
pyperclip          ==1.9.0
pyqr               ==0.0.1
PyQRCode           ==1.2.1
PyRect             ==0.2.0
PyScreeze          ==1.0.1
python-dateutil    ==2.9.0.post0
python-dotenv      ==1.1.0
pytweening         ==1.2.0
pytz               ==2025.1
pyunormalize       ==16.0.0
pywhatkit          ==5.4
pywin32            ==309
qrcode             ==8.0
queuelib           ==1.7.0
ramp               ==0.1.4
RapidFuzz          ==3.12.2
regex              ==2024.11.6
requests           ==2.32.3
rlp                ==4.1.0
rpcudp             ==5.0.0
rubicon-objc       ==0.5.0
semantic-version   ==2.10.0
setuptools         ==76.0.0
six                ==1.17.0
sniffio            ==1.3.1
soupsieve          ==2.6
sqlparse           ==0.5.3
tabulate           ==0.9.0
threadpoolctl      ==3.6.0
toolz              ==1.0.0
tqdm               ==4.67.1
types-requests     ==2.32.0.20250306
typing_extensions  ==4.12.2
typing-inspection  ==0.4.0
tzdata             ==2025.1
u-msgpack-python   ==2.8.0
urllib3            ==2.3.0
varint             ==1.0.2
w3lib              ==2.3.1
web3               ==7.10.0
websockets         ==13.1
Werkzeug           ==3.1.3
wget               ==3.2
wikipedia          ==1.4.0
yarl               ==1.18.3
z3                 ==0.2.0